Move methods from Basket Provider to BasketController.
